# lib/api-zod/src/generated/types/fuelCastMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FuelCastMetrics {
  observations: number;
  maeKgPerSecond: number;
  rmseKgPerSecond: number;
  r2: number;
  /** @nullable */
  mapePercent: number | null;
  /** @nullable */
  mapeUnavailableReason: string | null;
}