# lib/api-zod/src/generated/types/fuelCastScenarioFeatures.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FuelCastScenarioFeatures {
  /** @exclusiveMinimum 0 */
  referenceSpeedKnots: number;
  /** @exclusiveMinimum 0 */
  shipSpeedOverGroundAtReferenceSpeed: number;
  /** @exclusiveMinimum 0 */
  totalShaftPowerAtReferenceSpeed: number;
  /** @minimum 0 */
  windSpeed10m: number;
  /** @minimum 0 */
  waveHeight: number;
  /** @minimum 0 */
  oceanCurrentVelocity: number;
}