# lib/api-zod/src/generated/types/fuelCastTimeBlock.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FuelCastMetrics } from './fuelCastMetrics';

export type FuelCastTimeBlock = FuelCastMetrics & {
  vessel: string;
  sourceOrderStart: number;
  sourceOrderEnd: number;
};