# lib/api-zod/src/generated/types/fuelCastValidation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FuelCastMetrics } from './fuelCastMetrics';
import type { FuelCastTimeBlock } from './fuelCastTimeBlock';
import type { FuelCastValidationUnseenVesselFoldsItem } from './fuelCastValidationUnseenVesselFoldsItem';

export interface FuelCastValidation {
  available: boolean;
  modelAvailable: boolean;
  dataset: string;
  target: string;
  targetUnit: string;
  /** @nullable */
  observationsLoaded: number | null;
  /** @nullable */
  observationsUsable: number | null;
  /** @nullable */
  vesselCount: number | null;
  vessels: string[];
  features: string[];
  /** @nullable */
  model: string | null;
  /** @nullable */
  trainingObservations: number | null;
  /** @nullable */
  testObservations: number | null;
  timeBlockMetrics: FuelCastMetrics;
  timeBlocks: FuelCastTimeBlock[];
  unseenVesselMetrics: FuelCastMetrics;
  unseenVesselFolds: FuelCastValidationUnseenVesselFoldsItem[];
  physicsBaselineStatus: string;
  physicsBaselineReason: string;
  limitations: string[];
}