# lib/api-zod/src/generated/types/fuelScenarioInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FuelScenarioInput {
  fuelId: string;
  /** @minimum 0 */
  bunkerPriceInrPerTonne: number;
  /** @exclusiveMinimum 0 */
  densityKgPerM3: number;
  /** @minimum -10 */
  wellToTankKgCo2ePerKg: number;
  /** @minimum 0 */
  nonCo2TankToWakeKgCo2ePerKg: number;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  methaneSlipPercent: number | null;
  /**
     * @minLength 2
     * @maxLength 300
     */
  bunkerPriceSource: string;
  bunkerPriceAsOfDate: Date;
  /**
     * @minLength 2
     * @maxLength 300
     */
  wellToTankSource: string;
  availableAtBunkeringPort: boolean;
}