# lib/api-zod/src/generated/types/optimizeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FuelCastScenarioFeatures } from './fuelCastScenarioFeatures';
import type { FuelScenarioInput } from './fuelScenarioInput';
import type { ObjectiveWeights } from './objectiveWeights';
import type { OptimizeInputFuelEstimator } from './optimizeInputFuelEstimator';

export interface OptimizeInput {
  vesselId: string;
  /** @exclusiveMinimum 0 */
  distanceNm: number;
  /** @exclusiveMinimum 0 */
  cargoTonnes: number;
  /** @exclusiveMinimum 0 */
  actualDraftM: number;
  /** @exclusiveMinimum 0 */
  minSpeedKnots: number;
  /** @exclusiveMinimum 0 */
  maxSpeedKnots: number;
  /** @exclusiveMinimum 0 */
  speedStepKnots: number;
  /**
     * @minimum 0
     * @maximum 12
     */
  windBeaufort: number;
  /** @minimum 0 */
  waveHeightM: number;
  /**
     * @minimum 0
     * @maximum 180
     */
  waveRelativeDirectionDeg: number;
  /**
     * @minimum -10
     * @maximum 10
     */
  currentAlongTrackKnots: number;
  /** @minimum 0 */
  berthWindowStartHour: number;
  /** @exclusiveMinimum 0 */
  berthWindowEndHour: number;
  /** @minimum 0 */
  carbonPriceInrPerTonne: number;
  /**
     * @minLength 2
     * @maxLength 300
     */
  carbonPriceSource: string;
  /** @minimum 0 */
  portFeesInr: number;
  /**
     * @minLength 2
     * @maxLength 300
     */
  portFeesSource: string;
  objectiveWeights: ObjectiveWeights;
  /** @minItems 1 */
  fuels: FuelScenarioInput[];
  /**
     * @minimum 64
     * @maximum 4096
     */
  evaluationBudget: number;
  /**
     * @minimum 0
     * @maximum 2147483647
     */
  randomSeed: number;
  fuelEstimator?: OptimizeInputFuelEstimator;
  fuelCastFeatures?: FuelCastScenarioFeatures;
}