# lib/api-zod/src/generated/types/paretoPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParetoPoint {
  fuelId: string;
  fuelName: string;
  speedKnots: number;
  fuelMassTonnes: number;
  fuelVolumeM3: number;
  totalCostInr: number;
  fuelRateKgPerSecond: number;
  fuelEstimator: string;
  wellToWakeKgCo2e: number;
  tankToWakeKgCo2e: number;
  durationHours: number;
  engineLoadFraction: number;
  feasible: boolean;
}